@@ -1,31 +1,16 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-
-def appCompatVersion = '1.3.0-alpha02'
-def materialVersion = '1.3.0-alpha04'
-def multidexVersion = "2.0.1"
-def constraintLayoutVersion = '2.1.0-alpha1'
-def toothpickVersion = '1.1.3'
-def rxAndroidVersion = '2.0.2'
-def rxJavaVersion = '2.2.3'
-def retrofitVersion = '2.6.4'
-def okHttpVersion = '4.3.0'
-def picassoVersion = '2.71828'
-def circularImageViewVersion = '3.2.0'
-def junitVersion = '4.13.1'
-def mockitoCoreVersion = '2.27.0'
-def mockitoKotlinVersion = '1.5.0'
-def androidTestRunnerVersion = '1.3.1-alpha02'
-def espressoCoreVersion = '3.4.0-alpha02'
-def navVersion = "2.3.1"
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+    id 'kotlin-kapt'
+}
 
 android {
-    compileSdkVersion 28
+    namespace 'ru.mamykin.exchange'
+    compileSdkVersion 33
     defaultConfig {
         applicationId "ru.mamykin.exchange"
-        minSdkVersion 15
-        targetSdkVersion 28
+        minSdkVersion 21
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -39,12 +24,15 @@ android {
             }
         }
         compileOptions {
-            sourceCompatibility JavaVersion.VERSION_1_8
-            targetCompatibility JavaVersion.VERSION_1_8
+            sourceCompatibility JavaVersion.VERSION_11
+            targetCompatibility JavaVersion.VERSION_11
         }
         kotlinOptions {
-            jvmTarget = "1.8"
+            jvmTarget = "11"
         }
+    }
+    buildFeatures {
+        buildConfig true
     }
     buildTypes {
         release {
