@@ -1,11 +1,11 @@
 buildscript {
-    ext.kotlin_version = '1.4.20'
+    ext.kotlin_version = '2.1.0'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.1'
+        classpath 'com.android.tools.build:gradle:8.7.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
