@@ -8,8 +8,9 @@ import ru.mamykin.exchange.core.di.module.NetworkModule
 import ru.mamykin.exchange.core.di.module.SchedulersModule
 import toothpick.Toothpick
 import toothpick.configuration.Configuration
-import toothpick.registries.FactoryRegistryLocator
-import toothpick.registries.MemberInjectorRegistryLocator
+import ru.mamykin.exchange.BuildConfig
+// import toothpick.registries.FactoryRegistryLocator
+// import toothpick.registries.MemberInjectorRegistryLocator
 
 @Suppress("unused")
 class ConverterApp : Application() {
@@ -23,9 +24,9 @@ class ConverterApp : Application() {
         if (BuildConfig.DEBUG) {
             Toothpick.setConfiguration(Configuration.forDevelopment().preventMultipleRootScopes())
         } else {
-            Toothpick.setConfiguration(Configuration.forProduction().disableReflection())
-            FactoryRegistryLocator.setRootRegistry(FactoryRegistry())
-            MemberInjectorRegistryLocator.setRootRegistry(MemberInjectorRegistry())
+            // Toothpick.setConfiguration(Configuration.forProduction().disableReflection())
+            // FactoryRegistryLocator.setRootRegistry(FactoryRegistry())
+            // MemberInjectorRegistryLocator.setRootRegistry(MemberInjectorRegistry())
         }
         val appScope = Toothpick.openScope(Scopes.APP_SCOPE)
         appScope.installModules(
