@@ -11,7 +11,7 @@ import toothpick.Toothpick
 inline fun <reified T : ViewModel> Fragment.viewModel() = viewModels<ConverterViewModel> {
     object : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        override fun <T : ViewModel> create(modelClass: Class<T>): T {
             return Toothpick.openScopes(Scopes.APP_SCOPE, this@viewModel).getInstance(modelClass)
         }
     }
